+++ R/tar_cue_age.R
@@ -41,6 +41,7 @@ tar_cue_age <- function(
   command = TRUE,
   depend = TRUE,
   format = TRUE,
+  repository = TRUE,
   iteration = TRUE,
   file = TRUE
 ) {
@@ -51,6 +52,7 @@ tar_cue_age <- function(
     command = command,
     depend = depend,
     format = format,
+    repository = repository,
     iteration = iteration,
     file = file
   )