+++ R/tar_cue_age_raw.R
@@ -54,6 +54,7 @@ tar_cue_age_raw <- function(
   command = TRUE,
   depend = TRUE,
   format = TRUE,
+  repository = TRUE,
   iteration = TRUE,
   file = TRUE
 ) {
@@ -84,6 +85,7 @@ tar_cue_age_raw <- function(
     command = command,
     depend = depend,
     format = format,
+    repository = repository,
     iteration = iteration,
     file = file
   )