+++ R/tar_cue_force.R
@@ -56,6 +56,7 @@ tar_cue_force <- function(
   command = TRUE,
   depend = TRUE,
   format = TRUE,
+  repository = TRUE,
   iteration = TRUE,
   file = TRUE
 ) {
@@ -65,6 +66,7 @@ tar_cue_force <- function(
     command = command,
     depend = depend,
     format = format,
+    repository = repository,
     iteration = iteration,
     file = file
   )